server-app/event_marketplace/requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.2
anyio==4.4.0
apache-airflow==2.9.3
apache-airflow-providers-common-io==1.3.2
apache-airflow-providers-common-sql==1.14.2
apache-airflow-providers-fab==1.2.1
apache-airflow-providers-ftp==3.10.0
apache-airflow-providers-http==4.12.0
apache-airflow-providers-imap==3.6.1
apache-airflow-providers-smtp==1.7.1
apache-airflow-providers-sqlite==3.8.1
apispec==6.6.1
argcomplete==3.4.0
asgiref==3.8.1
astunparse==1.6.3
attrs==23.2.0
Babel==2.15.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
biopython==1.84
bitarray==2.9.2
blinker==1.7.0
bs4==0.0.2
cachelib==0.9.0
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.2
connexion==2.14.2
contourpy==1.2.1
cron-descriptor==1.4.3
croniter==2.0.7
cryptography==42.0.8
cycler==0.12.1
Deprecated==1.2.14
dill==0.3.8
distlib==0.3.8
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
docutils==0.21.2
email_validator==2.2.0
Faker==37.1.0
filelock==3.15.1
Flask==2.2.5
Flask-AppBuilder==4.5.0
Flask-Babel==2.0.0
Flask-Bcrypt==1.0.1
Flask-Caching==2.3.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.7.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-Script==2.0.6
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
flatbuffers==24.3.25
fonttools==4.53.0
fpdf==1.7.2
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.6.0
gast==0.6.0
geomet==0.2.1.post1
google-pasta==0.2.0
google-re2==1.1.20240702
googleapis-common-protos==1.63.2
gpt4all==2.8.2
greenlet==3.0.3
grpcio==1.64.1
gunicorn==22.0.0
h11==0.14.0
h5py==3.11.0
html5lib==1.1
http_client==0.1.22
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
inflection==0.5.1
iniconfig==2.0.0
intel-openmp==2021.4.0
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keras==3.4.1
kiwisolver==1.4.5
latexcodec==3.0.0
lazy-object-proxy==1.10.0
libclang==18.1.1
limits==3.13.0
linkify-it-py==2.0.3
lockfile==0.12.2
lxml==5.2.2
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
matplotlib==3.9.1
mdit-py-plugins==0.4.1
mdurl==0.1.2
methodtools==0.4.7
mkl==2021.4.0
ml-dtypes==0.3.2
more-itertools==10.3.0
mpmath==1.3.0
multidict==6.0.5
multitasking==0.0.11
mysql-connector-python==8.4.0
namex==0.0.8
neo4j==5.22.0
networkx==3.3
numpy==1.26.4
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opt-einsum==3.3.0
optree==0.11.0
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
pathspec==0.12.1
patsy==0.5.6
pdfreader==0.1.15
peewee==3.17.6
pendulum==3.0.0
pillow==10.3.0
platformdirs==4.2.2
pluggy==1.4.0
prison==0.2.1
protobuf==4.25.3
psutil==6.0.0
psycopg2-binary==2.9.9
py4j==0.10.9.7
pyasynchat==1.0.4
pyasyncore==1.0.4
pybtex==0.24.0
pycparser==2.22
pycryptodome==3.20.0
pycurl==7.45.4
pyftpdlib==2.0.1
Pygments==2.18.0
PyJWT==2.8.0
pymed==0.8.9
PyMySQL==1.1.1
pyparsing==3.1.2
PyPDF2==3.0.1
PySocks==1.7.1
pyspark==3.5.5
pytest==8.1.1
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-nmap==0.7.1
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2024.1
pywin32==308
PyYAML==6.0.1
referencing==0.35.1
reportlab==4.2.2
requests==2.32.3
requests-toolbelt==1.0.0
rest-framework-simplejwt==0.0.2
rfc3339-validator==0.1.4
rich==13.7.1
rich-argparse==1.5.2
rpds-py==0.19.0
scapy==2.6.1
scikit-learn==1.5.1
scipy==1.14.0
selenium==4.22.0
setproctitle==1.3.3
setuptools==70.0.0
shutup==0.2.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==1.4.52
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.0
statsmodels==0.14.3
sympy==1.12.1
tabulate==0.9.0
tbb==2021.12.0
tenacity==8.5.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==3.5.0
time-machine==2.14.2
tomli==2.0.1
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tqdm==4.66.4
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal_pathlib==0.2.2
urllib3==2.2.1
virtualenv==20.26.3
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wget==3.2
wheel==0.43.0
wirerope==0.4.7
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.1.2
xlwt==1.3.0
xmltodict==0.13.0
yarl==1.9.4
yfinance==0.2.43
yt-dlp==2024.11.18
zipp==3.19.2
